fix(fsm): RedisStorage с DefaultKeyBuilder(with_destiny=True) - обязателен для aiogram_dialog; errors-handler по канону + guard-тест; price -> Decimal; pyproject дополнен пинами (aiogram-dialog, dishka, redis, maxo, agent-framework), uv.lock пересобран

+++ concierge_bot/db/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     BigInteger,
@@ -44,7 +45,7 @@ class Service(Base):
     name: Mapped[str] = mapped_column(String(255))
     category: Mapped[str] = mapped_column(String(100))
     description: Mapped[str | None] = mapped_column(Text, nullable=True)
-    price: Mapped[float | None] = mapped_column(Numeric(10, 2), nullable=True)
+    price: Mapped[Decimal | None] = mapped_column(Numeric(10, 2), nullable=True)
     currency: Mapped[str] = mapped_column(String(3), server_default=text("'USD'"))
     is_active: Mapped[bool] = mapped_column(Boolean, server_default=true())
     created_at: Mapped[datetime] = mapped_column(server_default=func.now())
